+++ src/main/java/kitchenpos/order/dao/repository/JpaTableGroupRepository.java
@@ -2,7 +2,7 @@ package kitchenpos.order.dao.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.menu.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 
 public interface JpaTableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
