+++ src/test/java/kitchenpos/order/domain/TableGroupTest.java
@@ -9,8 +9,6 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import kitchenpos.menu.domain.TableGroup;
-
 public class TableGroupTest {
 
     @Test
